requirements.txt
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.3.0
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
croniter==1.3.4
dask==2022.4.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distributed==2022.4.1
docker==5.0.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
frozenlist==1.3.0
fsspec==2022.3.0
grpcio==1.43.0
HeapDict==1.0.1
idna==3.3
importlib-resources==5.7.1
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
locket==0.2.1
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-oneofschema==3.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.3
mypy-extensions==0.4.3
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathspec==0.9.0
peewee==3.14.10
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prefect==1.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py4j==0.10.9.3
pyarrow==7.0.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.8
pyrsistent==0.18.1
pyspark==3.2.1
pytest==7.1.1
python-box==6.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.1
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
ray==1.12.0
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tabulate==0.8.9
tblib==1.7.0
terminado==0.13.3
text-unidecode==1.3
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing-extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
widgetsnbextension==3.6.0
zict==2.1.0
zipp==3.8.0
